requirements.txt
pytest
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy